projet-e2-main/requirements.txt
absl-py==2.0.0
alembic==1.12.0
ansi2html==1.8.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.4
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.7.0
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.4
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.2
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.49.1
h5py==3.9.0
idna==3.4
imageai==3.0.3
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mock==5.1.0
mpmath==1.3.0
mysql-connector==2.2.9
mysqlclient==2.1.1
nest-asyncio==1.5.7
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==21.3
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
scikeras==0.11.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.1.0
zipp==3.17.0
